requirements.txt
ask-sdk-core
ask-sdk-model
ask-sdk-dynamodb-persistence-adapter
boto3
pytest